scraper/requirements.txt
gitdb==4.0.12
GitPython==3.1.45
lizard==1.19.0
pathspec==0.12.1
PyDriller==2.9
Pygments==2.19.2
pytz==2025.2
smmap==5.0.2
types-pytz==2025.2.0.20251108
